+++ plone/app/vocabularies/actions.py
@@ -1,5 +1,6 @@
 # -*- coding:utf-8 -*-
 from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone import PloneMessageFactory as _
 from zope.interface import implementer
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
@@ -26,3 +27,26 @@ class ActionCategoriesVocabulary(object):
 
 
 ActionCategoriesVocabularyFactory = ActionCategoriesVocabulary()
+
+
+@implementer(IVocabularyFactory)
+class PortalActionCategoriesVocabulary(object):
+    """Provides portal action categories vocabulary."""
+
+    def __call__(self, context):
+        portal_actions = getToolByName(context, 'portal_actions')
+        categories = portal_actions.objectIds()
+        categories.sort()
+        terms = []
+        for category in categories:
+            if category == 'controlpanel':
+                continue
+            title = portal_actions.get(category).title
+            if title:
+                title = _(title)
+            else:
+                title = category
+            terms.append(SimpleTerm(category, title=title))
+        return SimpleVocabulary(terms)
+
+PortalActionCategoriesVocabularyFactory = PortalActionCategoriesVocabulary()